Archives list
This adds a very simple listing of the archives that have been completed
(that have a WACZ file). There is some JavaScript that parses the CSV in
the GoogleSheet when the page is loaded. At the moment the links to view
the archive target ArchiveWeb.page, but we could use a local viewer
instead. I think that would better demonstrate how the web archive files
are not locked up in a particular service. If there is interest in
trying that I can submit an additional pull request.